Move get_resolution to _datastore.py
